refactor: nix netrics build
The Netrics build is by now more than similar enough to the generic
build that it doesn't warrant special support. Though this means special
configuration for netrics, this is trivial for traceneck, (as opposed to
less compatible tools); and, current deployment of netrics entails such
configuration, regardless, (to specify either NDT or Ookla speed-
testing).

As such, it should be sufficient to execute traceneck as follows under
netrics:

  traceneck -qo -

a.k.a.:

  traceneck --quiet --out-file=-

(i.e. quiet and writing to stdout)

or more fully:

  traceneck --quiet --out-file=- --tool=[ndt|ookla]

---

Otherwise:

* .dockerignore file re-added (missing since repo migration): recommended
  with Dockerfile COPY commands
* Dockerfile made to target GO version in use to avoid development
  environment issues
* Dockerfile made to use newer HEREDOC syntax with RUN commands
* `docker create` used to create inactive container in lieu of
  `run … sleep` in CI